Add files via upload
Analysis of Employee Attrition - The analysis states that among the total 1562 employees, the organisation consists of 48.96% males and 51.04% females. The average employee rating has been consistent over the years with 2.99 being the lowest in 2022 and 3.03 being the highest in the years 2016 and 2022. HR department has the highest number of employees (326) by followed by Quality Assurance (170) and IT (!62) while Customer Service has the least number of employees (139) but customer service department has performed the best in terms of average rating followed by sales and R&D. Lastly, in January of 2023, highest number of actions were taken upon employees with the major action being Hiring ( nearly 60%).